+++ VMath/PluckStringKS.h
@@ -7,11 +7,11 @@ private: /* data */
 	unsigned char pow = 2;
 
 public: /* constructor */
-	PluckStringKS(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit);
-	PluckStringKS(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit, unsigned char _pow);
+	PluckStringKS(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit);
+	PluckStringKS(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit, unsigned char _pow);
 
 protected: /* virtual */
-	bool synthesis(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit) override;
+	bool synthesis(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit) override;
 };
 
 class PluckStringEKS : public DigitalWaveguide
@@ -24,9 +24,9 @@ private: /* data */
 	double pos = 0.100;
 
 public: /* constructor */
-	PluckStringEKS(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit);
-	PluckStringEKS(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit, double _decay, double _damp, double _tune, double _mod, double _pos);
+	PluckStringEKS(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit);
+	PluckStringEKS(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit, double _decay, double _damp, double _tune, double _mod, double _pos);
 
 protected: /* virtual */
-	bool synthesis(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit) override;
+	bool synthesis(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit) override;
 };
