+++ VMath/WaveFunction.h
@@ -55,17 +55,17 @@ protected: /* parts */
     static bool isWaveHeader(const WaveHeader& _header);
     static bool isWaveFile(const std::string& _fname);
 
-    bool setWaveHeader(unsigned short _srate, unsigned short _sbit);
+    bool setWaveHeader(unsigned int _srate, unsigned short _sbit);
     bool setWaveData(const WaveData& _data);
 
-    static unsigned long long calcWaveDataSize(double _dura, unsigned short _srate);
+    static unsigned long long calcWaveDataSize(double _dura, unsigned int _srate);
 
 public: /* constructor */
     WaveFunction();
 
 public: /* public use */
     bool setWaveFunction(const WaveData& _data, const WaveHeader& _header);
-    bool setWaveFunction(const WaveData& _data, unsigned short _srate, unsigned short _sbit);
+    bool setWaveFunction(const WaveData& _data, unsigned int _srate, unsigned short _sbit);
     bool setWaveFunction(const WaveData& _data);
 
     bool exportWave(const std::string& _fname) const;
@@ -80,9 +80,9 @@ public: /* public use */
     WaveData getWaveData() const;
     short getWaveData(unsigned long long _tdx) const;
 
-    static WaveFunction sin(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit);
-    static WaveFunction sqr(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit, double _duty);
-    static WaveFunction tri(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit);
-    static WaveFunction saw(double _namp, double _freq, double _dura, unsigned short _srate, unsigned short _sbit, bool _reverse);
-    static WaveFunction ofs(double _ofs, double _dura, unsigned short _srate, unsigned short _sbit);
+    static WaveFunction sin(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit);
+    static WaveFunction sqr(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit, double _duty);
+    static WaveFunction tri(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit);
+    static WaveFunction saw(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit, bool _reverse);
+    static WaveFunction ofs(double _ofs, double _dura, unsigned int _srate, unsigned short _sbit);
 };
